Refactor frame code out into its own file.
