+++ engine/include/engine/frame_reader.h
@@ -5,6 +5,7 @@
 #include <vector>
 
 #include "absl/status/statusor.h"
+#include "engine/frame.h"
 
 extern "C" {
 #include <libavcodec/avcodec.h>
@@ -12,32 +13,6 @@ extern "C" {
 }  // extern "C"
 
 namespace viduce::engine {
-
-class Frame {
- public:
-  // Extra information about the stream that the frame belongs to.
-  struct StreamInfo {
-    int stream_index = -1;
-    AVMediaType media_type = AVMEDIA_TYPE_UNKNOWN;
-    AVCodecID codec_id = AV_CODEC_ID_NONE;
-  };
-
-  static std::unique_ptr<Frame> Create(AVStream* stream);
-
-  ~Frame() { av_frame_free(&frame_); }
-
-  AVFrame* frame() { return frame_; }
-
-  const StreamInfo& stream_info() const { return stream_info_; }
-
- private:
-  explicit Frame(const StreamInfo& stream_info) : stream_info_(stream_info) {
-    frame_ = av_frame_alloc();
-  }
-
-  AVFrame* frame_;
-  const StreamInfo stream_info_;
-};
 
 class FrameReader {
  public:
